Добавлена статистика, удаление+редактирование команды, просмотр заявок

+++ src/main/java/com/searchteam/bot/entity/User.java
@@ -33,7 +33,7 @@ public class User {
 
     private Integer currentProjectChoice;
 
-    private Integer currentTeamChoice;
+    private Long currentTeamChoice;
 
     private Long currentRequestChoice;
